backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
exceptiongroup==1.2.2
fastapi==0.115.3
h11==0.14.0
idna==3.10
netifaces==0.10.6
paho-mqtt==2.1.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.0
typing_extensions==4.12.2
uvicorn==0.32.0
websockets==13.1
